requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
pydantic>=2.4
pydantic-settings>=2.0
numpy>=1.25
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers>=4.43
datasets>=2.16
peft>=0.11.0
accelerate>=0.30
trl>=0.9.6
jinja2>=3.1

